refactor(lit): レイアウト統一
