+++ packages/lit/index.html
@@ -6,13 +6,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>@kotice/lit-web-components-sandbox</title>
     <link rel="stylesheet" href="./src/assets/globals.css" />
-    <script type="module" src="/src/examples/my-example-element.ts"></script>
+    <script type="module" src="/src/my-app.ts"></script>
     <script type="module" src="/src/components/lit/my-search-input.ts"></script>
   </head>
   <body>
-    <my-element>
-      <h1>@kotice/lit-web-components-sandbox</h1>
+    <my-app>
       <my-search-input></my-search-input>
-    </my-element>
+    </my-app>
   </body>
 </html>
