+++ packages/lit/package.json
@@ -11,7 +11,9 @@
     }
   },
   "types": "./dist/@types/index.d.ts",
-  "files": ["dist"],
+  "files": [
+    "dist"
+  ],
   "scripts": {
     "dev": "vite",
     "build": "vite build && tsc"
