+++ packages/vue/src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import './assets/globals.css'
 import App from './App.vue'
+import './assets/globals.css'
 
 createApp(App).mount('#app')
